Code refactoring [x3]

+++ SCP999/Interfaces/IAbility.cs
@@ -4,7 +4,7 @@ public interface IAbility
     string Name { get; }
     string Description { get; }
     int KeyId { get; }
-    int Cooldown { get; }
+    float Cooldown { get; }
     void Register();
     void Unregister();
 }
